@@ -5,26 +5,36 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "hyram"
 authors = [
-    {name = "Brian D. Ehrhart"}, 
-    {name = "Cianan Sims"}, 
-    {name = "Ethan S. Hecht"}, 
-    {name = "Benjamin B. Schroeder"}, 
-    {name = "Benjamin R. Liu"}, 
-    {name = "Katrina M. Groth"}, 
-    {name = "John T. Reynolds"}, 
-    {name = "Gregory W. Walkup"}
+    {name = "Erin E. Carrier"},
+    {name = "Michael C. Devin"},
+    {name = "Brian D. Ehrhart"},
+    {name = "Isaac W. Ekoto"},
+    {name = "Katrina M. Groth"},
+    {name = "Ethan S. Hecht"},
+    {name = "Benjamin R. Liu"},
+    {name = "Alice Muna"},
+    {name = "John T. Reynolds"},
+    {name = "Benjamin B. Schroeder"},
+    {name = "Cianan Sims"},
+    {name = "Greg Walkup"}
+]
+maintainers = [
+    {name = "Michael C. Devin"},
+    {name = "Brian D. Ehrhart"},
+    {name = "Ethan S. Hecht"},
+    {name = "Benjamin R. Liu"},
+    {name = "Cianan Sims"}
 ]
 description = "Hydrogen Plus Other Alternative Fuels Risk Assessment Models Python Package"
 readme = "./src/hyram/README.md"
-requires-python = ">3.6, <3.11"
+requires-python = ">=3.9,<3.13"
 keywords = ["hydrogen", "methane", "propane", "flame", "plume", "jet", "overpressure", "quantitative risk assessment", "risk"]
-license = {text = "GPL-3.0-only"}
+license = "GPL-3.0-only"
 classifiers = [
     "Development Status :: 5 - Production/Stable",
-    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
     "Programming Language :: Python :: 3 :: Only"
 ]
-dependencies = ["matplotlib", "numpy", "scipy", "coolprop >= 6.3"]
+dependencies = ["matplotlib >= 3.8", "numpy", "scipy", "coolprop >= 6.3", "contourpy"]
 dynamic = ["version"]
 
 [project.urls]
