@@ -1,5 +1,5 @@
 """
-Copyright 2015-2024 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
+Copyright 2015-2025 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
 Under the terms of Contract DE-NA0003525 with NTESS, the U.S. Government retains certain rights in this software.
 
 You should have received a copy of the GNU General Public License along with HyRAM+.
@@ -9,7 +9,7 @@ If not, see https://www.gnu.org/licenses/.
 from ._jet import Jet
 from ._indoor_release import IndoorRelease
 from ._flame import Flame
-from ._comps import Fluid, Orifice, Source, Enclosure, Vent
+from ._comps import Fluid, Orifice, NozzleFlow, Source, Enclosure, Vent
 from ._unconfined_overpressure import BST_method, TNT_method, Bauwens_method
 from ._fuel_props import FuelProperties
 from . import api
