@@ -1,5 +1,5 @@
 """
-Copyright 2015-2024 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
+Copyright 2015-2025 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
 Under the terms of Contract DE-NA0003525 with NTESS, the U.S. Government retains certain rights in this software.
 
 You should have received a copy of the GNU General Public License along with HyRAM+.
@@ -25,7 +25,7 @@ class Fluid:
         """
         class used to describe a fluid (usually a gas)
         two of four (T, P, rho, phase) are needed to fully define the fluid
-        
+
         Parameters
         ----------
         species: string or dictionary
@@ -67,7 +67,7 @@ class Fluid:
             therm = CoolPropWrapper(self.species)
 
         params = therm._init_fluid_params({'D':rho, 'P':P, 'T':T, 'phase':phase})
-        
+
         self.rho, self.P, self.T, self.phase = params['D'], params['P'], params['T'], params['phase']
         self.therm = therm
         self.v = v
@@ -92,18 +92,18 @@ class Orifice:
     def __init__(self, d, Cd=1):
         '''
         class used to describe a circular orifice
-        
+
         future versions may be expanded to give effective area for other shapes
-        
+
         Parameters
         ----------
         d - orifice diameter (m)
         Cd - discharge coefficient to account for non-plug flow (always <=1, assumed to be 1 for plug flow)
-        
+
         Contains
         --------
         d - diameter (m)
-        Cd- discharge coefficient 
+        Cd- discharge coefficient
         A - effective area (m^2)
         '''
         self.d, self.Cd, self.A = d, Cd, np.pi / 4 * d ** 2
@@ -111,51 +111,41 @@ class Orifice:
     def __repr__(self):
         return 'orifice\n%s\ndiameter = %.2f mm\ndischarge coefficient = %.2f' % (30 * '-', self.d * 1e3, self.Cd)
 
-    def mdot(self, fluid):
+class NozzleFlow:
+    def __init__(self, upstream_fluid, orifice, downstream_P, mdot = None, suppressWarnings = True):
         '''
-        mass flow rate through orifice of a fluid object
+        Class used to calculate isentropic flow through an orifice
         
         Parameters
         ----------
-        rho - density (kg/m^3)
-        v - velocity (m/s)
+        upstream_fluid - fluid object upstream of orifice
+        orifice - Orifice object through which upstream fluid flows
+        downstream_P - downstream pressure (Pa)
+        mdot - mass flow rate (if fluid is unchoked, it is recommended to provide) [kg/s]
+        suppressWarning - boolean whether to suppress warnings or not
         
         Returns
         -------
-        mdot - mass flow rate (kg/s)
+        fluid - fluid at the nozzle with state (pressure, temperature, density) and velocity
         '''
-        return fluid.rho * fluid.v * self.A * self.Cd
-
-    def flow(self, upstream_fluid, downstream_P = 101325, mdot = None, suppressWarnings = True):
-        '''
-        Returns the fluid in a flow restriction, for given upstream conditions 
-        and downstream pressure.  Isentropic expansion.
-        
-        Parameters
-        ----------
-        upstream_fluid - upstream fluid with therm object, as well as P, T, rho, v
-        downstream_P - float: downstream pressure (Pa)
-        mdot - mass flow rate - only used for unchoked flow
-        
-        Returns
-        -------
-        Fluid object containing T, P, rho, v at the throat (orifice)
-        '''           
+        self.upstream_fluid, self._orifice, self.downstream_P = upstream_fluid, orifice, downstream_P
         h0 = upstream_fluid.therm.get_property('H0', P=upstream_fluid.P, D=upstream_fluid.rho, v=upstream_fluid.v)
         if upstream_fluid.v > 0:
             s0 = upstream_fluid.therm.get_property('S', H = h0, D = np.round(upstream_fluid.rho, 12))
         else: #LH2 simulations were giving weird results when calculating entropy from enthalpy
             s0 = upstream_fluid.therm.get_property('S', D = upstream_fluid.rho, T = upstream_fluid.T)
-        
+
         fluid = copy.copy(upstream_fluid)
+
         if fluid.P < downstream_P:
             raise ValueError('Downstream pressure is higher than upstream pressure.  Nonphysical.')
         if fluid.P == downstream_P:
             if mdot is not None:
-                v = mdot/(self.Cd*fluid.rho*self.A)
+                v = mdot/(orifice.Cd*fluid.rho*orifice.A)
                 fluid.update(v=v)
-                fluid._choked = False
-                return fluid
+                self.choked = False
+                self.fluid = fluid
+                return
             else:
                 raise ValueError('Downstream pressure is the same as upstream pressure.  Need to specify mass flow rate (mdot).')
 
@@ -166,22 +156,44 @@ class Orifice:
         P = optimize.minimize_scalar(negflux, bounds = (downstream_P, upstream_fluid.P), method = 'bounded')['x']
         h, rho = fluid.therm.get_property(['H', 'D'], P = P, S = s0)
         fluid.update(rho = rho, P = P, v = np.sqrt(2 * (h0 - h)))
-        if P - downstream_P > .01: 
+        if P - downstream_P > .01:
             if mdot is not None:
                 if not suppressWarnings:
                     warnings.warn('Fluid choked. Ignoring mdot specification and using choked flow calculation.', category=PhysicsWarning)
-            fluid._choked = True
+            self.choked = True
         else: 
-            fluid._choked = False
+            self.choked = False
             if mdot is None:
                 if not suppressWarnings:
                     warnings.warn('Fluid unchoked. Provide a mass flow rate.', category=PhysicsWarning)
             else:
-                v = mdot/(self.Cd*rho*self.A)
+                v = mdot/(orifice.Cd*rho*orifice.A)
                 fluid.update(rho=rho, P=(P*(P - downstream_P > .01) + 
                                          downstream_P*(P - downstream_P <= .01)), v=v)
-        return fluid
+        self.fluid = fluid
 
+    @property
+    def orifice(self):
+        return self._orifice
+
+    @orifice.setter
+    def orifice(self, new_orifice):
+        if self.choked: #choked - P, rho, T, v remain constant at orifice, only ._orifice needs to be updated so mass flow rate is properly calculated
+            self._orifice = new_orifice
+        else: #unchoked - previously specified mass flow rate through orifice assumed constant; fluid velocity must also be updated
+            self.fluid.v = self.mdot/(self.fluid.rho*new_orifice.Cd*new_orifice.A)
+            self._orifice = new_orifice
+    
+    @property
+    def mdot(self):
+        '''
+        mass flow rate through the nozzle
+        
+        Returns
+        -------
+        mdot - mass flow rate (kg/s)
+        '''
+        return self.fluid.rho * self.fluid.v * self.orifice.A * self.orifice.Cd
 
 class Source(object):
     """
@@ -197,7 +209,7 @@ class Source(object):
     def __init__(self, V, fluid):
         '''
         Initializes source based on the volume and the fluid object in the source (tank)
-        
+
         Parameters
         ----------
         V: float
@@ -218,14 +230,14 @@ class Source(object):
     def fromMass(cls, m, fluid):
         '''
         Initilization method based on the mass and the fluid object in the source (tank)
-        
+
         Parameters
         ----------
         m: float
             mass of source (tank) (kg)
         fluid: object
             fluid object in the source (tank)
-        
+
         Returns
         -------
         source: object
@@ -237,9 +249,9 @@ class Source(object):
     @classmethod
     def fromMass_Vol(cls, species, m, V, T=None, P=None):
         '''
-        Initilization method based on the mass, volume, and either the temperature or pressure of the fluid 
+        Initilization method based on the mass, volume, and either the temperature or pressure of the fluid
         in the source (tank).
-        
+
         Parameters
         ----------
         species: string
@@ -254,7 +266,7 @@ class Source(object):
             temperature (K)
         P: float (optional)
             pressure (Pa)
-        
+
         Returns
         -------
         source: object
@@ -270,9 +282,10 @@ class Source(object):
             return None
         return cls(V, fluid)
 
-    def _blowdown_gov_eqns(self, t, ind_vars, Vol, orifice, heat_flux, ambient_P):
-        '''governing equations for energy balance on a tank (https://doi.org/10.1016/j.ijhydene.2011.12.047)
-        
+    def _blowdown_gov_eqns(self, t, ind_vars, Vol, orifice, heat_flux, ambient_P, interp):
+        '''
+        governing equations for energy balance on a tank (https://doi.org/10.1016/j.ijhydene.2011.12.047)
+
         Parameters
         ----------
         t - time (s)
@@ -280,7 +293,8 @@ class Source(object):
         Vol - float, volume of tank (m^3)
         orifice - orifice object
         heat_flux - float, heat flow into tank (W)
-        
+        interp - dictionary of interpolating functions for isentropic blowdown or None if not isentropic
+
         Returns
         -------
         [dm_dt, du_dt] = array of [d(mass)/dt (kg/s), d(internal energy)/dt (J/kg-s)]
@@ -290,112 +304,148 @@ class Source(object):
         m, U = ind_vars
         m, U = float(m), float(U)
         rho = m / Vol
-        T = therm.get_property('T', U=U, D=np.round(rho, 10))
-        fluid = copy.copy(self.fluid)
-        fluid.update(T=T, rho=rho)
-        throat = orifice.flow(fluid, ambient_P)
-        h = therm.get_property('H', T=fluid.T, D=fluid.rho)
-        dm_dt = -orifice.mdot(throat)
+        if interp is not None:
+            def err_U(P):
+                return interp['U'](P) - U
+            P = optimize.root(err_U, ambient_P)['x'][0]
+            T, rho_interp, h = [float(interp[k](P)) for k in ['T', 'D', 'H']]
+            fluid = copy.copy(self.fluid)
+            fluid.update(T=T, rho=rho)
+        else:
+            T = therm.get_property('T', U=U, D=np.round(rho, 10))
+            fluid = copy.copy(self.fluid)
+            fluid.update(T=T, rho=rho)
+            h = therm.get_property('H', T=fluid.T, D=fluid.rho)
+
+        if fluid.P < ambient_P:
+            return np.array([0, 0])
+        throat = NozzleFlow(fluid, orifice, ambient_P)
+
+        dm_dt = -throat.mdot
         du_dt = 1 / m * (heat_flux + (h - U) * dm_dt)
         return np.array([dm_dt, du_dt])
-   
+
     def empty(self, orifice, ambient_P = 101325,
-              heat_flux = 0, nmax = 1000, 
+              heat_flux = 0, nmax = 1000,
               m_empty = 1e-6, p_empty_percent = 0.01,
-              t_empty = None,
-              timeout = 10):
+              t_empty = 3600*24,
+              timeout = 600):
         '''
-        integrates the governing equations for an energy balance on a tank 
-        
+        integrates the governing equations for an energy balance on a tank
+
         Parameters
         ----------
-        orifice - orifice object through which the source is emptying
-        ambient_P - ambient pressure into which leak occurs (Pa)
-        heat_flux - Heat flow (W) into tank.  Assumed to be 0 (adiabatic)
-        nmax - maximum number of iterations
-        m_empty - mass when considered empty (kg)
-        p_empty_percent - percent of ambient pressure when considered empty
-        timeout - computer run time (minutes) until you give up on the integration
-        t_empty - if given, empties for a set amount of time (sec)
-        
+        orifice : Orifice
+            Orifice object through which the source is emptying
+        ambient_P : float, optional
+            Ambient pressure into which leak occurs (Pa), defaults to 1 atm
+        heat_flux : float, optional
+            Heat flow (W) into tank.  Assumed to be 0 (adiabatic)
+        nmax : int, optional
+            Maximum number of iterations, defaults to 1000
+        m_empty : float, optional
+            Mass when considered empty (kg), defaults to 1e-6
+        p_empty_percent : float, optional
+            Percent of ambient pressure when considered empty, defaults to 0.01%
+        t_empty : int, optional
+            maximum blowdown time (sec), defaults to 24 hours
+        timeout : int, optional
+            maximum computer run time (sec) until the integration terminates, defaults to 600s
+
         Returns
         -------
         tuple of (mdot, fluid_list, t, solution_array) =
-                 (list of mass flow rates (kg/s), list of fluid objects at each time step, 
+                 (list of mass flow rates (kg/s), list of fluid objects at each time step,
                   array of times (s), 2D array of [mass, internal energy] at each time step)
         '''
-        start_time = time.time()
         therm = self.fluid.therm
         m0 = self.m
         volume = self.V
-        A_orifice = orifice.A * orifice.Cd
-        T0, rho0, P = self.fluid.T, self.fluid.rho, self.fluid.P
-        u0 = therm.get_property('U', T=T0, D=rho0)
-        r = integrate.ode(self._blowdown_gov_eqns).set_integrator('dopri5')
-        r.set_initial_value([m0, u0]).set_f_params(volume, orifice, heat_flux, ambient_P)
-        throat = orifice.flow(self.fluid, ambient_P)
-        if t_empty is None:
-            mdot0 = orifice.mdot(throat)
-            dt = m0 / mdot0
+
+        is_isentropic = (heat_flux == 0)
+        if is_isentropic and 'interp' not in self.fluid.therm.__dict__:
+            therm.make_isentropic_interpolating_functions(self.fluid, ambient_P, self.fluid.P)
+        if is_isentropic:
+            interp = self.fluid.therm.interp
+            u0 = float(interp['U'](self.fluid.P))
         else:
-            dt = t_empty
-        times, fluid_list, mdot, sol = [], [], [], []
+            interp = None
+            u0 = therm.get_property('U', T=self.fluid.T, D=self.fluid.rho)
 
-        def solout(t, y):
-            if len(times) > 0:
-                if times[-1] == t:
-                    return
-            rho = y[0] / volume
-            T = therm.get_property('T', U=y[1], D=np.round(rho, 10))
-            fluid = copy.copy(self.fluid)
-            fluid.update(T=T, rho=rho)
-            throat = orifice.flow(fluid, ambient_P)
+        # Define ending state events
+        start_time = time.time()
+        n_steps = 0
+        ambient_P_threshold = ambient_P * (1 + p_empty_percent/100)
+
+        def fluid_from_m_U(m, U):
+            '''fluid in tank, given mass (m) and internal energy (U)'''
+            rho = m / volume
+            if interp is not None:
+                def err_U(P):
+                    return interp['U'](P) - U
+                P = optimize.root(err_U, ambient_P)['x'][0]
+                T, rho_interp, h = [float(interp[k](P)) for k in ['T', 'D', 'H']]
+                fluid = copy.copy(self.fluid)
+                fluid.update(T=T, rho=rho)
+            else:
+                T = therm.get_property('T', U=U, D=np.round(rho, 10))
+                fluid = copy.copy(self.fluid)
+                fluid.update(T=T, rho=rho)
+            return fluid
+
+        # Set function attribute 'terminal' to end integration when one of these events is met
+        def eventAttr():
+            def decorator(f):
+                f.terminal = True
+                return f
+            return decorator
+
+        @eventAttr()
+        def time_out(*args):
+            current = time.time() - start_time
+            return timeout - current
+
+        @eventAttr()
+        def max_steps(*args):
+            nonlocal n_steps
+            n_steps += 1
+            return nmax - n_steps
+
+        @eventAttr()
+        def residual_mass(*args):
+            return m_empty - args[1][0]
+
+        @eventAttr()
+        def residual_pressure(*args):
+            tank_P = fluid_from_m_U(*args[1]).P
+            return tank_P - ambient_P_threshold
+
+        sol = integrate.solve_ivp(self._blowdown_gov_eqns, (0, t_empty), [m0, u0],
+                                  method='BDF',
+                                  events=(time_out, max_steps, residual_mass, residual_pressure),
+                                  args=(volume, orifice, heat_flux, ambient_P, interp))
+
+        if sol.status == 0:
+            warnings.warn("Blowdown not complete after time limit")
+
+        times = list(sol['t'])
+        tank_sol = sol['y'] # list of [[mass], [internal energy]]
+        tank_states = tank_sol.T  # turn into list of tuples [(mass, internal energy)]
+
+        # Generate lists of fluids and mdot values
+        fluid_list, mdot = [], []
+        for state in tank_states:
+            fluid = fluid_from_m_U(*state)
+            throat = NozzleFlow(fluid, orifice, ambient_P)
             fluid_list.append(fluid)
-            mdot.append(orifice.mdot(throat))
-            times.append(t)
-            sol.append([y[0], y[1]])
+            mdot.append(throat.mdot)
 
-        r.set_solout(solout)
-        nsteps = 0
-        while (True):
-            nsteps += 1
-            try:
-                r.integrate(r.t + dt)
-            except:
-                for i in range(100):
-                    try:
-                        # need to back up to last successful integration point
-                        j = np.argmin(np.abs(np.array(times) - r.t))
-                        times = times[:j + 1]
-                        fluid_list = fluid_list[:j + 1]
-                        mdot = mdot[:j + 1]
-                        sol = sol[:j + 1]
-                        dt /= 10
-                        try:
-                            r.integrate(r.t + dt)
-                            break
-                        except:
-                            pass
-                    except:
-                        print(i, 'unable to advance past time %.5f s with remaining mass of %.3f g, P = %.1f Pa' %
-                              (r.t, r.y[0] * 1000, fluid_list[-1].P))
-                        return mdot, fluid_list, times, np.array(sol).T
-            if time.time() - start_time > timeout*60:
-                print(f'timeout of {timeout} min reached')
-                break
-            if nsteps > nmax:
-                print('max steps reached')
-                break
-            if not r.successful():
-                break
-            if t_empty is not None:
-                if r.t >= t_empty:
-                    break
-            elif (r.y[0] < m_empty or fluid_list[-1].P < (1 + p_empty_percent / 100) * ambient_P):
-                break
-        self.mdot, self.fluid_list, self.ts, self.sol = mdot, fluid_list, times, np.array(sol).T 
-        return mdot, fluid_list, times, np.array(sol).T
-    
+        if is_isentropic:
+            delattr(self.fluid.therm, 'interp')
+
+        self.mdot, self.fluid_list, self.ts, self.sol = mdot, fluid_list, times, tank_sol
+        return mdot, fluid_list, times, tank_sol
+
     def plot_time_to_empty(self):
         try:
             x = self.sol[0]
@@ -424,11 +474,11 @@ class Enclosure:
     def __init__(self, H, A, H_release, ceiling_vent, floor_vent, Xwall=np.inf):
         '''
         Describes the enclosure
-        
+
         Parameters
         ----------
         H : encosure height (m)
-        A : area of floor and ceiling (m^2) 
+        A : area of floor and ceiling (m^2)
         H_release : height of release (m)
         ceiling_vent : vent class containing vent information for ceiling vent
         floor_vent : vent class containing vent information for floor vent
@@ -447,7 +497,7 @@ class Vent:
     def __init__(self, A, H, Cd=1, vol_flow_rate=0):
         '''
         Describes the vent
-        
+
         Parameters
         ----------
         A : vent cross-sectional area (m^2)
