@@ -1,5 +1,5 @@
 """
-Copyright 2015-2024 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
+Copyright 2015-2025 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
 Under the terms of Contract DE-NA0003525 with NTESS, the U.S. Government retains certain rights in this software.
 
 You should have received a copy of the GNU General Public License along with HyRAM+.
@@ -9,24 +9,22 @@ If not, see https://www.gnu.org/licenses/.
 import copy
 
 import matplotlib.pyplot as plt
+from contourpy import contour_generator
 import numpy as np
 from scipy import integrate, optimize
 import scipy.constants as const
 
 from ._fuel_props import FuelProperties
 from ._notional_nozzle import NotionalNozzle
+from ._comps import NozzleFlow
 from ._plots import plot_contour
-
-########################################################################
-# TODO: untested for alternate fuels
-########################################################################
 
 
 class DevelopingFlow:
     def __init__(self, fluid, orifice, ambient, mdot=None,
                  theta0=0, x0=0, y0=0,
                  lam=1.16, betaA=0.28,
-                 nn_conserve_momentum=True, nn_T='solve_energy', 
+                 nn_conserve_momentum=True, nn_T='solve_energy',
                  T_establish_min=-1,
                  suppressWarnings=False, verbose=False):
         '''
@@ -44,29 +42,32 @@ class DevelopingFlow:
         self._nn_conserve_momentum = nn_conserve_momentum
         S0 = 0 # S always starts at 0. x and y may start somewhere else.
         Y0 = 1 # pure fluid
-
+        
         # Orifice flow
         self._orifice = orifice
         if self.verbose:
             print('solving for orifice flow... ', end='')
-        self.fluid_orifice = self.orifice.flow(fluid, ambient.P, mdot, suppressWarnings) # plug node at orifice exit
-        self.mass_flow_rate = self.orifice.mdot(self.fluid_orifice)
+        
+        self.orifice_flow = NozzleFlow(fluid, self.orifice, ambient.P, mdot, suppressWarnings) # plug node at orifice exit
+        self.fluid_orifice = self.orifice_flow.fluid
+        
+        self.mass_flow_rate = self.orifice_flow.mdot
         if self.verbose:
             print('done')
         self.d0 = orifice.d
         self.orifice_node = PlugNode(orifice.d*np.sqrt(orifice.Cd), self.fluid_orifice.v, self.fluid_orifice.rho,# diameter scaled to account for Cd
                                      1, self.fluid_orifice.T, theta0, x0, y0, S0)
 
-        # Underexpanded jet (if needed: gets fluid to atmospheric pressure)
-        self.fluid_exp, self.orifice_exp = self._expand(orifice, ambient, nn_T, nn_conserve_momentum)
-
+        # Underexpanded jet (gets fluid to atmospheric pressure (if needed, otherwise retains same properties))
+        self.fluid_exp, self.orifice_exp = self._expand(ambient, nn_T, nn_conserve_momentum)
+ 
         # Initial entrainment and heating (if needed: warms fluid to good T for thermodynamics)
-        self.expanded_plug_node = self._dev_plug(self.fluid_exp, self.orifice_exp, ambient, Y0, theta0, x0, y0, S0, 
+        self.expanded_plug_node = self._dev_plug(self.fluid_exp, self.orifice_exp, ambient, Y0, theta0, x0, y0, S0,
                                                  T_establish_min, betaA)
 
         # Develop into established Gaussian profile from plug flow
         self.initial_node = self.expanded_plug_node.establish(ambient, self.fluid_exp, lam)
-    
+
     @property
     def lam(self):
         return self._lam
@@ -100,10 +101,9 @@ class DevelopingFlow:
         self.expanded_plug_node.x = x0 + S_out*np.cos(val)
         self.expanded_plug_node.y = y0 + S_out*np.sin(val)
         self.expanded_plug_node.theta = val
-        SE = 6.2*self.expanded_plug_node.d
+        SE = self.initial_node.S - self.expanded_plug_node.S
         self.initial_node.x = self.expanded_plug_node.x + SE*np.cos(val)
         self.initial_node.y = self.expanded_plug_node.y + SE*np.sin(val)
-        self.initial_node.S = self.expanded_plug_node.S + SE
         self.initial_node.theta = val
         self._theta0 = val
 
@@ -111,19 +111,18 @@ class DevelopingFlow:
     def orifice(self):
         return self._orifice
 
-    #todo: could probably improve orifice.setter - can likely redo _expand with less overhead - notional nozzle models may need to be modified
+    #todo: could probably improve orifice.setter - can likely redo _expand with less overhead - if notional nozzle models were modified
     @orifice.setter
     def orifice(self, new_orifice):
-        if self.fluid_orifice._choked: #choked - P, rho, T, v remain constant at orifice (fluid_orifice is constant)
-            self.mass_flow_rate = new_orifice.mdot(self.fluid_orifice)
-        else: #unchoked - mass flow rate through orifice remains constant
-            self.fluid_orifice.v = self.mass_flow_rate/(self.fluid_orifice.rho*new_orifice.Cd*new_orifice.A)
+        self.orifice_flow.orifice = new_orifice
+        self.mass_flow_rate = self.orifice_flow.mdot
+        self.fluid_orifice = self.orifice_flow.fluid
         self._orifice = new_orifice
         self.d0 = new_orifice.d
         self.orifice_node = PlugNode(self.d0*np.sqrt(new_orifice.Cd), self.fluid_orifice.v, self.fluid_orifice.rho,# diameter scaled to account for Cd
-                                     1, self.fluid_orifice.T, self.theta0, self.x0, self.y0, self.S0)
+                                     1, self.fluid_orifice.T, self.theta0, self.x0, self.y0, 0)
         # Underexpanded jet (if needed: gets fluid to atmospheric pressure)
-        self.fluid_exp, self.orifice_exp = self._expand(new_orifice, self.ambient, self._nn_T, self._nn_conserve_momentum)
+        self.fluid_exp, self.orifice_exp = self._expand(self.ambient, self._nn_T, self._nn_conserve_momentum)
 
         # Initial entrainment and heating (if needed: warms fluid to good T for thermodynamics)
         self.expanded_plug_node = self._dev_plug(self.fluid_exp, self.orifice_exp, self.ambient,
@@ -134,22 +133,22 @@ class DevelopingFlow:
         self.initial_node = self.expanded_plug_node.establish(self.ambient, self.fluid_exp, self.lam)
 
 
-    def _expand(self, orifice, ambient, nn_T, nn_conserve_momentum):
+    def _expand(self, ambient, nn_T, nn_conserve_momentum):
         '''
         expands an underexpanded jet, if needed
         '''
         if self.fluid_orifice.P > ambient.P: # use notional nozzle model
             if self.verbose:
                 print('solving for notional nozzle... ', end='')
-            nn = NotionalNozzle(self.fluid_orifice, orifice, ambient)
+            nn = NotionalNozzle(self.orifice_flow, ambient)
             g, o = nn.calculate(nn_T, nn_conserve_momentum)
             if self.verbose:
                 print('done.')
         else:
-            g, o = self.fluid_orifice, orifice
+            g, o = self.fluid_orifice, self.orifice
         return g, o
 
-    def _dev_plug(self, fluid, orifice, ambient, Y0, theta0, x0, y0, S0, 
+    def _dev_plug(self, fluid, orifice, ambient, Y0, theta0, x0, y0, S0,
                   T_establish_min, betaA):
         '''
         zone of initial entrainment and heating (if needed)
@@ -161,7 +160,7 @@ class DevelopingFlow:
             if self.verbose:
                 print('solving for zone of initial entrainment and heating... ', end='')
             air_out = copy.copy(ambient)
-            air_out.update(T = T_establish_min, P = air_out.P) 
+            air_out.update(T = T_establish_min, P = air_out.P)
             fluid_out = copy.copy(fluid)
             fluid_out.update(T = T_establish_min, P = fluid_out.P)
             mdot_in = orifice.mdot(fluid)
@@ -180,10 +179,10 @@ class DevelopingFlow:
             rho_out = 1/((1-Y)/air_out.rho + Y/fluid_out.rho)
             mdot_out = mdot_in + (1-Y)/Y*mdot_in
             v_out = fluid.v*mdot_in/mdot_out
-            E = betaA*np.sqrt(orifice.mdot(fluid)*fluid.v/ambient.rho) 
+            E = betaA*np.sqrt(orifice.mdot(fluid)*fluid.v/ambient.rho)
             S_out = (1-Y)*(mdot_out - mdot_in)/E/ambient.rho
             d_out = np.sqrt(mdot_out/(rho_out*v_out)*4/np.pi)
-            plug_node = PlugNode(d_out, v_out, rho_out, Y, fluid_out.T, theta0, 
+            plug_node = PlugNode(d_out, v_out, rho_out, Y, fluid_out.T, theta0,
                                  x0 + S_out*np.cos(theta0), y0 + S_out*np.sin(theta0), S0 + S_out)
             if self.verbose:
                 print('done.')
@@ -192,7 +191,7 @@ class DevelopingFlow:
 class PlugNode:
     def __init__(self, d, v, rho, Y, T, theta, x = 0, y = 0, S = 0):
         '''Plug flow node
-        
+
         Parameters
         ----------
         d: float, diameter (m)
@@ -207,11 +206,11 @@ class PlugNode:
         self.d, self.v, self.rho, self.Y = d, v, rho, Y
         self.theta, self.x, self.y, self.S = theta, x, y, S
         self.T = T
-    
+
     def Froude(self, rho_amb):
-        '''Froude number'''
+        '''Densimetric Froude number'''
         return self.v/np.sqrt(const.g*self.d*abs(rho_amb - self.rho)/self.rho) #ESH - 7/23/20 - added abs to work with negatively buoyant jets
-    
+
     def establish(self, ambient, fluid, lam):
         '''
         returns the gaussian node after flow establishment
@@ -225,7 +224,7 @@ class PlugNode:
         MW_clE = 1 / (Y_clE / fluid.therm.MW + (1 - Y_clE) / ambient.therm.MW)
         h_amb = ambient.T*cp_ambient
         cp_s3 = self.Y * cp_gas + (1 - self.Y) * cp_ambient # this is just cp_gas - using self.Y screws it up
-        
+
         T_s3 = self.T
         # TODO: Use ThermoProps to calculate enthalpy
         h_s3 = cp_s3 * T_s3
@@ -234,11 +233,11 @@ class PlugNode:
         T_clE = h_clE / cp_clE
 
         rho_clE = ambient.P*MW_clE/(const.R*T_clE)
-        #######################################################################################
-        # Note: correlation below is for Fr**2 > 40, which is pretty much always true for these 
-        # gases - could add Froude number correlation if desired
-        #######################################################################################
-        SE = 6.2*self.d
+        Fr_den = self.Froude(ambient.rho)
+        SE = self.d*((6.2                  ) if (Fr_den**2 >= 40) else
+                     (3.9 + 0.057*Fr_den**2) if (5<=Fr_den**2<40) else
+                     (2.075+0.425*Fr_den**2) if (1<=Fr_den**2<5 ) else
+                     (0                    ))
         xE = self.x + SE*np.cos(self.theta)
         yE = self.y + SE*np.sin(self.theta)
         SE = self.S + SE
@@ -248,8 +247,8 @@ class PlugNode:
 class GaussianNode:
     def __init__(self, B, v_cl, rho_cl, Y_cl, theta, x = 0, y = 0, S = 0):
         '''
-        Describes the properties of a node 
-        
+        Describes the properties of a node
+
         Parameters
         ----------
         B: float
@@ -271,19 +270,19 @@ class GaussianNode:
         '''
         self.B, self.v_cl, self.rho_cl, self.Y_cl = B, v_cl, rho_cl, Y_cl
         self.S, self.theta, self.x, self.y = S, theta, x, y
-    
+
     @property
     def conditions(self):
-        return np.array([self.v_cl, self.B, self.rho_cl, self.Y_cl, self.theta, self.x, self.y])    
+        return np.array([self.v_cl, self.B, self.rho_cl, self.Y_cl, self.theta, self.x, self.y])
 
     def entrainment(self, Emom, rho_amb, alpha_buoy, alpha):
-        FrL = self.v_cl**2*self.rho_cl/(const.g*self.B*abs(rho_amb-self.rho_cl)) # ESH: added absolute value 07/23/20 - for negatively buoyant jets, this was a negative number - might need justification
-        E_buoy = alpha_buoy/FrL*(2*const.pi*self.v_cl*self.B)*np.sin(self.theta)  # m**2/s
+        Fr_L = self.v_cl**2*self.rho_cl/(const.g*self.B*abs(rho_amb-self.rho_cl)) # ESH: added absolute value 07/23/20 - for negatively buoyant jets, this was a negative number - might need justification
+        E_buoy = alpha_buoy/Fr_L*(2*const.pi*self.v_cl*self.B)*np.sin(self.theta)  # m**2/s
         E = Emom + E_buoy
         alphatest = E/(2*const.pi*self.v_cl*self.B)
         if alphatest > alpha:
             E = alpha*2*const.pi*self.B*self.v_cl
-        return E        
+        return E
 
 
 class Jet:
@@ -298,9 +297,9 @@ class Jet:
                  developing_flow = None,
                  suppressWarnings=False, verbose=False):
         '''
-        Class for solving for a 2D jet. 
+        Class for solving for a 2D jet.
         If fluid pressure is <= 2 x ambient pressure, use subsonic initialization (specify mdot).
-               
+
         Parameters
         ----------
         fluid: fluid object
@@ -316,7 +315,7 @@ class Jet:
         x0: float, optional
             horizontal starting location (m)
         y0: float, optional
-            vertical starting location (m)     
+            vertical starting location (m)
         lam : float, optional
             Relative spreading ratio of concentration to velocity
         betaA : float, optional
@@ -325,7 +324,7 @@ class Jet:
             whether notional nozzle model should conserve mass and momentum, or mass only,
             together with nn_T determines which notional nozzle model to use (see below)
         nn_T: string, optional
-            either 'solve_energy', 'Tthroat' or specified temperature (stagnation temperature) 
+            either 'solve_energy', 'Tthroat' or specified temperature (stagnation temperature)
             with nn_conserve_momentum leads to one of the following notional nozzle models:
             YuceilOtugen - conserve_momentum = True, T = 'solve_energy'
             EwanMoodie - conserve_momentum = False, T = 'Tthroat'
@@ -333,7 +332,7 @@ class Jet:
             Birch2 - conserve_momentum = True, T = T0
             Molkov - conserve_momentum = False, T = 'solve_energy'
         T_establish_min: float, optional
-            minimum temperature (K) at the zone of flow establishment, if specified, will implement the 
+            minimum temperature (K) at the zone of flow establishment, if specified, will implement the
             zone of initial entrainment and heating if temperature after notional nozzle model is lower
         Ymin: float, optional
             minimum mass fraction to integrate to (default is about 1 mol%)
@@ -359,7 +358,7 @@ class Jet:
             whether to display warnings about fluid being under-/over-specified in DevelopingFlow object
         verbose: boolean, optional
             whether to include print statements about the model actions
-        There are up to 4 engineering models that give initial conditions to an 
+        There are up to 4 engineering models that give initial conditions to an
         integral model:
         1) flow through the orifice - choked if pressure above critical pressure, assumed
            to be plug flow at volumetric flow rate if not
@@ -388,10 +387,10 @@ class Jet:
         self.mass_flow_rate = self.developing_flow.mass_flow_rate
 
         # Set up some entrainment parameters:
-        Fr = self.developing_flow.expanded_plug_node.Froude(ambient.rho)
-        self.Fr = Fr
-        if Fr < 268:
-            self._alpha_buoy = 17.313-0.11665*Fr+(2.0771e-4)*Fr**2 
+        Fr_den = self.developing_flow.expanded_plug_node.Froude(ambient.rho)
+        self.Fr_den = Fr_den
+        if Fr_den < 268:
+            self._alpha_buoy = 17.313-0.11665*Fr_den+(2.0771e-4)*Fr_den**2
         else:
             self._alpha_buoy = 0.97
         expanded_plug_node = self.developing_flow.expanded_plug_node
@@ -442,7 +441,7 @@ class Jet:
                    developing_flow = developing_flow,
                    suppressWarnings=suppressWarnings, verbose=verbose)
 
-    def solve(self, Ymin = 7e-4, dS = None, Smax = np.inf, 
+    def solve(self, Ymin = 7e-4, dS = None, Smax = np.inf,
               max_steps = 5000, tol = 1e-8,
               alpha = 0.082, Yamb = 0, numB = 5, numpts = 500):
         '''
@@ -452,28 +451,28 @@ class Jet:
             print('integrating... ', end='')
 
         if dS is None and Smax == np.inf:
-            dS = 500*self.developing_flow.expanded_plug_node.d # somewhat arbitrary - solver will add points anyway 
-                                        # may integrate past Ymin 
+            dS = 500*self.developing_flow.expanded_plug_node.d # somewhat arbitrary - solver will add points anyway
+                                        # may integrate past Ymin
         elif dS is None:
             dS = Smax
 
         r = integrate.ode(self._govEqns).set_f_params(alpha, Yamb, numB, numpts)
         r.set_integrator('dopri5', atol = tol, rtol = tol)
-        
+
         T, Y = [], []
         def solout(t, y):
             T.append(t)
             Y.append(np.array(y))
         r.set_solout(solout)
         r.set_initial_value(self.initial_node.conditions, self.initial_node.S)
-        
+
         i = 0
         while r.successful() and r.y[3] > Ymin and i < max_steps and r.t < Smax:
             r.integrate(r.t + dS)
             i += 1
-            
+
         Y = np.array(Y)
-        
+
         for key, val in zip(['V_cl', 'B', 'rho_cl', 'Y_cl', 'theta', 'x', 'y'], Y.T):
             self.__dict__[key] = val
         self.__dict__['S'] = np.array(T)
@@ -482,19 +481,19 @@ class Jet:
         MW_cl  = MW_air*MW_fluid/(self.Y_cl*(MW_air-MW_fluid) + MW_fluid)
         self.X_cl = self.Y_cl*MW_cl/MW_fluid
         self.T_cl = self.ambient.P*MW_cl/(const.R*self.rho_cl)
-        
+
         if self.verbose:
             print('done.')
 
         return self
-    
+
     def _govEqns(self, S, ind_vars, alpha = 0.082, Yamb = 0, numB = 5, numpts = 500):
         '''
-        Governing equations for a plume, written in terms of d/dS of (V_cl, B, rho_cl, Y_cl, 
+        Governing equations for a plume, written in terms of d/dS of (V_cl, B, rho_cl, Y_cl,
         theta, x, and y).
-        
-        A matrix solution to the continuity, x-momentum, y-momentum, species, and energy 
-        equations solves for d/dS of the dependent variables V_cl, B, rho_cl, Y_cl,  and Theta.  
+
+        A matrix solution to the continuity, x-momentum, y-momentum, species, and energy
+        equations solves for d/dS of the dependent variables V_cl, B, rho_cl, Y_cl,  and Theta.
         Numerically integrated to infinity = numB * B(S) using numpts discrete points.
         '''
         # break independent variables out of ind_vars, then put them into node_in
@@ -515,7 +514,7 @@ class Jet:
         V       = V_cl*np.exp(-(r**2)/(B**2))
         dVdS = np.array([V/V_cl,                                                 #d/dS(V_cl)
                          2*V*r**2/B**3,                                          #d/dS(B)
-                         zero,                                                   #d/dS(rho_cl) 
+                         zero,                                                   #d/dS(rho_cl)
                          zero,                                                   #d/dS(Y_cl)
                          zero])                                                  #d/dS(theta)
         rho     = (rho_cl - rho_amb)*np.exp(-(r**2)/((lam*B)**2))+rho_amb
@@ -528,32 +527,10 @@ class Jet:
                          zero])                                                     #d/dS(theta)
         MW      = MW_air*MW_fluid/(Y*(MW_air - MW_fluid) + MW_fluid)
         dMWdS   = (MW*(MW_air - MW_fluid)/(MW_fluid*(Y-1) - MW_air*Y))*dYdS
-        Cp      = Y*(Cp_fluid - Cp_air) + Cp_air 
+        Cp      = Y*(Cp_fluid - Cp_air) + Cp_air
         dCpdS   = (Cp_fluid - Cp_air)*dYdS
         rhoh    = Pamb/const.R*MW*Cp
         drhohdS = Pamb/const.R*(MW*dCpdS + Cp*dMWdS)
-        ##########################################################
-        # TODO: integrating the energy equation without involving Cp - not sure what the isssue is in the code below
-        # drhodS  = np.array([zero,                                                #d/dS(V_cl)
-                            # -2*r**2*(rho_amb - rho_cl)*np.exp(r**2/(lam*B)**2),  #d/dS(B)
-                            # lam**2*B**3*np.exp(r**2/(lam*B)**2),                 #d/dS(rho_cl)
-                            # zero,                                                #d/dS(Y_cl)
-                            # zero                                                 #d/dS(theta)
-                            # ])*1./(lam**2*B**3)
-        # # TODO: remove ideal gas assumption here (low priority)
-        # T = Pamb*MW/(const.R*rho)
-        # dTdS = Pamb/(const.R*rho)*dMWdS - Pamb*MW/(const.R*rho**2)*drhodS
-        # h_amb = self._h_amb(T) #self.ambient.therm.h(T = T, P = Pamb)
-        # d_h_amb_dT = self._dh_amb_dT(T)
-        # h_fluid = self._h_fluid(T)
-        # d_h_fluid_dT = self._dh_fluid_dT(T)
-        # h = Y*h_fluid - Y*h_amb + h_amb
-        # dhdS = (h_fluid - h_amb)*dYdS + Y*(d_h_fluid_dT - d_h_amb_dT)*dTdS + d_h_amb_dT*dTdS
-        
-        # rhoh = rho*h
-        # drhohdS = h*drhodS + rho*dhdS
-        # #########################################################
-
 
         # governing equations:
         LHScont = np.array([(lam**2*rho_cl + rho_amb)*B**2,                        #d/dS(V_cl)
@@ -562,40 +539,40 @@ class Jet:
                             0,                                                    #d/dS(Y_cl)
                             0])*const.pi/(lam**2 + 1)                             #d/dS(theta)
         RHScont = rho_amb*E
-        
+
         LHSxmom = np.array([(2*lam**2*rho_cl+rho_amb)*B**2*V_cl*np.cos(theta),     #d/dS(V_cl)
                             (2*lam**2*rho_cl+rho_amb)*B*V_cl**2*np.cos(theta),     #d/dS(B)
                             lam**2*B**2*V_cl**2*np.cos(theta),                     #d/dS(rho_cl)
                             0,                                                    #d/dS(Y_cl)
                             -(2*lam**2*rho_cl+rho_amb)*(B*V_cl)**2*np.sin(theta)/2 #d/dS(theta)
-                            ])*const.pi/(2*lam**2+1)        
+                            ])*const.pi/(2*lam**2+1)
         RHSxmom = 0
-        
+
         LHSymom = np.array([(2*lam**2*rho_cl+rho_amb)*B**2*V_cl*np.sin(theta),     #d/dS(V_cl)
                             (2*lam**2*rho_cl+rho_amb)*B*V_cl**2*np.sin(theta),     #d/dS(B)
                             lam**2*B**2*V_cl**2*np.sin(theta),                     #d/dS(rho_cl)
                             0,                                                    #d/dS(Y_cl)
                             (2*lam**2*rho_cl+rho_amb)*(B*V_cl)**2*np.cos(theta)/2  #d/dS(theta)
-                            ])*const.pi/(2*lam**2+1)                                 
+                            ])*const.pi/(2*lam**2+1)
         RHSymom = -const.pi*lam**2*const.g*(rho_cl - rho_amb)*B**2
-        
+
         LHSspec = np.array([B*Y_cl*rho_cl,                                         #d/dS(V_cl)
                             2*V_cl*Y_cl*rho_cl,                                    #d/dS(B)
                             B*V_cl*Y_cl,                                           #d/dS(rho_cl)
                             B*V_cl*rho_cl,                                         #d/dS(Y_cl)
                             0,                                                    #d/dS(theta)
-                            ])*const.pi*lam**2*B/(lam**2 + 1)                
+                            ])*const.pi*lam**2*B/(lam**2 + 1)
         RHSspec = Yamb*RHScont
-        
-        LHSener = 2*const.pi*integrate.trapz(V*drhohdS*r + rhoh*dVdS*r, r)
+
+        LHSener = 2*const.pi*integrate.trapezoid(V*drhohdS*r + rhoh*dVdS*r, r)
         LHSener += [const.pi/(6*lam**2 + 2)*(3*lam**2*rho_cl+rho_amb)*B**2*V_cl**2, #d/dS(V_cl)
                     const.pi/(9*lam**2 + 3)*(3*lam**2*rho_cl+rho_amb)*V_cl**3*B,    #d/dS(B)
                     const.pi/(6*lam**2 + 2)*lam**2*B**2*V_cl**3,                    #d/dS(rho_cl)
                     0,                                                             #d/dS(Y_cl)
                     0]                                                             #d/dS(theta)
-        
+
         RHSener = h_amb0*RHScont
-        
+
         LHS = np.array([LHScont,
                         LHSxmom,
                         LHSymom,
@@ -607,14 +584,14 @@ class Jet:
                         RHSymom,
                         RHSspec,
                         RHSener])
-        
+
         dz = np.append(np.linalg.solve(LHS,RHS), np.array([np.cos(theta), np.sin(theta)]), axis = 0)
-        
+
         return dz
-    
+
     def reshape(self, enclosure, showPlot = False):
         '''
-        reshapes the plume to turn upwards, should it hit the enclosure wall, 
+        reshapes the plume to turn upwards, should it hit the enclosure wall,
         and crops it so it stops at the ceiling
         '''
         if np.any(self.x > enclosure.Xwall):
@@ -639,7 +616,7 @@ class Jet:
         '''
         Calculates the amount of mass in the plume
         that is within the flammability limits
-        
+
         Parameters
         ----------
         plume : Plume class
@@ -653,7 +630,7 @@ class Jet:
         Hmax : float, optional
             Maximum height for integration
             Default is infinity (np.inf)
-        
+
         Outputs
         -------
         mass : float
@@ -670,44 +647,58 @@ class Jet:
         Ylean = X_lean * MW_fluid / (X_lean * MW_fluid + (1 - X_lean) * MW_air)
         Yrich = X_rich * MW_fluid / (X_rich * MW_fluid + (1 - X_rich) * MW_air)
 
-        # Trim the plume down to below Hmax:
         S = np.copy(self.S)
         Y_cl = np.copy(self.Y_cl)
         B = np.copy(self.B)
         rho_cl = np.copy(self.rho_cl)
-        if np.all(self.y > Hmax) or np.all(self.Y_cl < Ylean) or np.all(self.Y_cl > Yrich): # no flammable mass 
+
+        if np.all(self.y > Hmax) or np.all(self.Y_cl < Ylean): # no flammable mass
             return 0
-        elif np.any(self.y > Hmax): # trim arrays to Hmax
-            S = np.append(S[np.argwhere(self.y < Hmax)].T[0], np.interp(Hmax, self.y, S))
-            Y_cl = np.append(Y_cl[np.argwhere(self.y < Hmax)].T[0], np.interp(Hmax, self.y, Y_cl))
-            B = np.append(B[np.argwhere(self.y < Hmax)].T[0], np.interp(Hmax, self.y, B))
-            rho_cl = np.append(rho_cl[np.argwhere(self.y < Hmax)].T[0], np.interp(Hmax, self.y, rho_cl))
-                
-        def rhoY(r, i):
-            'mass fraction and density at radius: r, for plume at node: i'
-            rho = (rho_cl[i]-self.ambient.rho)*np.exp(-r**2/(self.lam*B[i]**2))+self.ambient.rho
-            Y = rho_cl[i]*Y_cl[i]*np.exp(-r**2/(self.lam*B[i])**2)/rho
-            return rho, Y
+
+        # Trim the plume down to below Hmax:
+        def trim_below(Hmax, y_vals, cl_vals):
+            return np.append(cl_vals[np.argwhere(y_vals < Hmax)].T[0], np.interp(Hmax, y_vals, cl_vals))
+        if np.any(self.y > Hmax):
+            S = trim_below(Hmax, self.y, S)
+            Y_cl = trim_below(Hmax, self.y, Y_cl)
+            B = trim_below(Hmax, self.y, B)
+            rho_cl = trim_below(Hmax, self.y, rho_cl)
+
+        def rho(r, i):
+            '''density at radius: r, for plume at node: i'''
+            return (rho_cl[i] - self.ambient.rho) * np.exp(-r**2 / (self.lam * B[i]**2)) + self.ambient.rho
+        def Y(r, i):
+            '''mass fraction at radius: r, for plume at node: i'''
+            return rho_cl[i] * Y_cl[i] * np.exp(-r**2 / (self.lam * B[i])**2) / rho(r, i)
 
         Slean = np.interp(Ylean, Y_cl[::-1], S[::-1])
-        Srich = np.interp(Yrich, Y_cl[::-1], S[::-1])
-        ivals = slice(np.argmax(Y_cl <= Yrich), np.argmax(Y_cl <= Ylean))
+        ivals = slice(0, np.argmax(Y_cl <= Ylean))
+        def trim_cl_lean(var):
+            return np.append(var[ivals], np.interp(Slean, S, var))
+        Y_cl = trim_cl_lean(Y_cl)
+        B = trim_cl_lean(B)
+        rho_cl = trim_cl_lean(rho_cl)
+        S = trim_cl_lean(S)
 
-        Y_cl, B, rho_cl, S = [np.append(np.append(np.interp(Srich, S, var), var[ivals]), np.interp(Slean, S, var)) for var in [Y_cl, B, rho_cl, S]]
-        # radius of flammable concentration at each node:
-        r_lean = np.array([0 if rhoY(0, i)[1] <= Ylean else 
-                           optimize.brentq(lambda r: rhoY(r, i)[1] - Ylean, 0, 100*B[i])
-                           for i in range(len(S))])
-        r_rich = np.array([0 if rhoY(0, i)[1] <= Yrich else 
-                           optimize.brentq(lambda r: rhoY(r, i)[1] - Yrich, 0, 100*B[i])
-                           for i in range(len(S))])
-
+        def rval_at_Yval(Yval):
+            if Y(0, i) <= Yval:
+                rval = 0
+            else:
+                rval = optimize.brentq(lambda r: Y(r, i) - Yval, 0, 100*B[i])
+            return rval
+        def mass_per_cylindrical_shell(r):
+            return Y(r, i) * rho(r, i) * 2 * const.pi * r
+        mass_per_len = []
         # integrate to find the mass/length at each node
-        mass_per_len = np.array([integrate.quad(lambda r: np.prod(rhoY(r, i))*2*const.pi*r, 
-                                 r_r, r_l)[0] for i, r_r, r_l in zip(range(len(S)), r_rich, r_lean)])
+        for i in range(len(S)):
+            r_lean_i = rval_at_Yval(Ylean)
+            r_rich_i = rval_at_Yval(Yrich)
+            mass_per_len_i = integrate.quad(mass_per_cylindrical_shell, r_rich_i, r_lean_i)[0]
+            mass_per_len.append(mass_per_len_i)
+
         # integrate to find the total mass
-        return integrate.trapz(mass_per_len, S)
-    
+        return integrate.trapezoid(mass_per_len, S)
+
     @property
     def _contourdata(self):
         """
@@ -728,31 +719,31 @@ class Jet:
             temperatures
         """
         iS = np.arange(len(self.S))
-        
+
         poshalf = np.logspace(-5, np.log10(3*np.max(self.B)))
         r = np.concatenate((-1 * poshalf[::-1], [0], poshalf))
-        
+
         r, iS = np.meshgrid(r, iS)
         B = self.B[iS]
         rho_cl = self.rho_cl[iS]
         Y_cl = self.Y_cl[iS]
-        V_cl = self.V_cl[iS]                    
-        
+        V_cl = self.V_cl[iS]
+
         rho_amb, Tamb, Pamb = self.ambient.rho, self.ambient.T, self.ambient.P
         MW_fluid, MW_air = self.fluid.therm.MW, self.ambient.therm.MW
-        
+
         rho = rho_amb + (rho_cl - rho_amb)*np.exp(-r**2/self.lam**2/B**2)
         Y   = Y_cl*rho_cl*np.exp(-(r**2)/((self.lam*B)**2))/rho
         MW  = MW_air*MW_fluid/(Y*(MW_air-MW_fluid) + MW_fluid)
-        
+
         X = Y*MW/MW_fluid
         v = V_cl*np.exp(-(r**2)/(B**2))
         T = Pamb*MW/(const.R*rho)
-        
+
         x = self.x[iS] + r*np.sin(self.theta[iS])
         y = self.y[iS] - r*np.cos(self.theta[iS])
         return x, y, X, Y, v, T
-    
+
     def _radial_profile(self, distance, ind_var = 'Y', nB = 3):
         '''
         Returns radial profile at a certain distance along the jet
@@ -882,22 +873,24 @@ class Jet:
                 contour2: {...}
             }
         """
-        fracs = [fracs] if type(fracs) in [int, float] else fracs
-        contours = np.sort(fracs)
+        contours = [fracs] if type(fracs) in [int, float] else fracs
         x_values, y_values, molefrac_values, __, __, __ = self._contourdata
-        cs = plt.contour(x_values, y_values, molefrac_values, levels=contours)
+        cg = contour_generator(x_values, y_values, molefrac_values)
         distances = {}
-        for contour, clc in zip(contours, cs.collections):
-            for path in clc.get_paths():
-                xs, ys = path.vertices.transpose()
 
-                xmin, xmax = min(xs), max(xs)
-                xmin = 0 if xmin > 0 and xmax > 0 else xmin
-                xmax = 0 if xmin < 0 and xmax < 0 else xmax
+        for contour in contours:
+            try:
+                xs, ys = cg.lines(contour)[0].T
+            except:
+                raise ValueError(f"There doesn't seem to be a contour at {contour}, try a value between {molefrac_values.min()} and {molefrac_values.max()}.")
 
-                ymin, ymax = min(ys), max(ys)
-                ymin = 0 if ymin > 0 and ymax > 0 else ymin
-                ymax = 0 if ymin < 0 and ymax < 0 else ymax
+            xmin, xmax = min(xs), max(xs)
+            xmin = 0 if xmin > 0 and xmax > 0 else xmin
+            xmax = 0 if xmin < 0 and xmax < 0 else xmax
 
-                distances[contour] = [(xmin, xmax), (ymin, ymax)]
+            ymin, ymax = min(ys), max(ys)
+            ymin = 0 if ymin > 0 and ymax > 0 else ymin
+            ymax = 0 if ymin < 0 and ymax < 0 else ymax
+
+            distances[contour] = [(xmin, xmax), (ymin, ymax)]
         return distances
