@@ -1,5 +1,5 @@
 """
-Copyright 2015-2024 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
+Copyright 2015-2025 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
 Under the terms of Contract DE-NA0003525 with NTESS, the U.S. Government retains certain rights in this software.
 
 You should have received a copy of the GNU General Public License along with HyRAM+.
@@ -21,6 +21,7 @@ def calc_probability_not_occur(probability):
     """
     Calculates complementary probability of a given event probability
     """
+    check_probability_value(probability)
     probability_not_occur = 1 - probability
     return probability_not_occur
 
