@@ -1,5 +1,5 @@
 """
-Copyright 2015-2024 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
+Copyright 2015-2025 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
 Under the terms of Contract DE-NA0003525 with NTESS, the U.S. Government retains certain rights in this software.
 
 You should have received a copy of the GNU General Public License along with HyRAM+.
@@ -8,6 +8,8 @@ If not, see https://www.gnu.org/licenses/.
 
 import unittest
 
+import numpy as np
+
 from hyram.qra import risk
 
 
@@ -15,13 +17,6 @@ class TestRiskMetricCalcs(unittest.TestCase):
     """
     Test basic calculation of risk metrics
     """
-    def test_calc_pll(self):
-        frequency = 1.12e-5  # per year
-        consequence = 0.1  # fatalties
-        pll = risk.calc_pll(frequency, consequence)
-        # Hand-calculation of above numbers
-        self.assertAlmostEqual(pll, 1.12e-6)
-
     def test_calc_far(self):
         pll = 2.5e-5  # fatalties/year
         total_occupants = 9
@@ -36,6 +31,11 @@ class TestRiskMetricCalcs(unittest.TestCase):
         # Hand-calculation of above numbers
         self.assertAlmostEqual(far, 0)
 
+    def test_zero_far_for_zero_occupants(self):
+        pll = 0.1
+        far = risk.calc_far(pll=pll, total_occupants=0)
+        self.assertEqual(far, 0)
+
     def test_calc_air(self):
         far = 0.03  # fatalties per 10^8 person*hours
         exposed_hours_per_year = 2000
@@ -49,8 +49,8 @@ class TestScenarioRiskCalcs(unittest.TestCase):
     Test calculation of risk (PLL) for all scenarios
     """
     def test_calc_all_plls(self):
-        frequencies = [1e-5, 2e-5]
-        fatalities = [0.1, 0.2]
+        frequencies = np.array([1e-5, 2e-5])
+        fatalities = np.array([0.1, 0.2])
         plls = risk.calc_all_plls(frequencies, fatalities)
         # Hand-calculation of above numbers
         hand_calc_plls = [1e-6, 4e-6]
@@ -58,7 +58,7 @@ class TestScenarioRiskCalcs(unittest.TestCase):
             self.assertAlmostEqual(calc_pll, test_pll)
 
     def test_calc_risk_contributions(self):
-        plls = [1e-6, 4e-6]
+        plls = np.array([1e-6, 4e-6])
         total_pll, pll_contributions = risk.calc_risk_contributions(plls)
         # Hand-calculation of above numbers
         self.assertAlmostEqual(total_pll, 5e-6)
@@ -67,7 +67,7 @@ class TestScenarioRiskCalcs(unittest.TestCase):
             self.assertAlmostEqual(calc_value, test_value)
 
     def test_calc_risk_contributions_handles_zeros(self):
-        plls = [0, 0]
+        plls = np.array([0, 0])
         total_pll, pll_contributions = risk.calc_risk_contributions(plls)
         # Hand-calculation of above numbers
         self.assertAlmostEqual(total_pll, 0)
