@@ -1,5 +1,5 @@
 """
-Copyright 2015-2024 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
+Copyright 2015-2025 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
 Under the terms of Contract DE-NA0003525 with NTESS, the U.S. Government retains certain rights in this software.
 
 You should have received a copy of the GNU General Public License along with HyRAM+.
@@ -13,13 +13,9 @@ import json
 from scipy.constants import pi
 
 from hyram.phys import Fluid, Orifice, Flame
-from hyram.utilities import misc_utils
 
 from . import utils
 
-"""
-NOTE: if running from IDE like pycharm, make sure cwd is hyram/ and not hyram/tests.
-"""
 
 # Flags to enable command line output, pyplots, and text file output
 VERBOSE = False
@@ -28,9 +24,8 @@ CREATE_OUTPUT = False
 
 # Absolute paths to input/output data
 DATA_LOC = os.path.join(os.path.dirname(__file__), 'data')
-OUTPUT_LOC = os.path.join(misc_utils.get_temp_folder(), 'validation-cryo')
-OUTPUT_FILE = os.path.join(OUTPUT_LOC, 'Validation.txt')
 LIMITS_FILE = os.path.join(DATA_LOC, 'Limits-cryo.json')
+OUTPUT_LOC = os.path.join('out', 'validation-cryo')
 
 
 class Test_LabCryoData(unittest.TestCase):
@@ -80,7 +75,7 @@ class Test_LabCryoData(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='m',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -109,7 +104,7 @@ class Test_LabCryoData(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='kW/m^2',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -138,7 +133,7 @@ class Test_LabCryoData(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='kW/m^2',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -167,7 +162,7 @@ class Test_LabCryoData(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='kW/m^2',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -196,7 +191,7 @@ class Test_LabCryoData(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='kW/m^2',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -225,7 +220,7 @@ class Test_LabCryoData(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='kW/m^2',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
