@@ -1,5 +1,5 @@
 """
-Copyright 2015-2024 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
+Copyright 2015-2025 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
 Under the terms of Contract DE-NA0003525 with NTESS, the U.S. Government retains certain rights in this software.
 
 You should have received a copy of the GNU General Public License along with HyRAM+.
@@ -13,13 +13,9 @@ import json
 import numpy as np
 
 from hyram.phys import Fluid, Orifice, Jet
-from hyram.utilities import misc_utils
 
 from . import utils
 
-"""
-NOTE: if running from IDE like pycharm, make sure cwd is hyram/ and not hyram/tests.
-"""
 
 # Flags to enable command line output, pyplots, and text file output
 VERBOSE = False
@@ -28,9 +24,8 @@ CREATE_OUTPUT = False
 
 # Absolute paths to input/output data
 DATA_LOC = os.path.join(os.path.dirname(__file__), 'data')
-OUTPUT_LOC = os.path.join(misc_utils.get_temp_folder(), 'validation-cryo_concentrations')
-OUTPUT_FILE = os.path.join(OUTPUT_LOC, 'Validation.txt')
 LIMITS_FILE = os.path.join(DATA_LOC, 'Limits-cryo_concentrations.json')
+OUTPUT_LOC = os.path.join('out', 'validation-cryo_concentrations')
 
 
 class Test_Xiao_2010(unittest.TestCase):
@@ -77,7 +72,7 @@ class Test_Xiao_2010(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -114,7 +109,7 @@ class Test_Xiao_2010(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -151,7 +146,7 @@ class Test_Xiao_2010(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -188,7 +183,7 @@ class Test_Xiao_2010(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -264,7 +259,7 @@ class Test_Friedrich_Dist1(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -301,7 +296,7 @@ class Test_Friedrich_Dist1(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -338,7 +333,7 @@ class Test_Friedrich_Dist1(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -375,7 +370,7 @@ class Test_Friedrich_Dist1(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -412,7 +407,7 @@ class Test_Friedrich_Dist1(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -449,7 +444,7 @@ class Test_Friedrich_Dist1(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -524,7 +519,7 @@ class Test_Friedrich_Dist2(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -561,7 +556,7 @@ class Test_Friedrich_Dist2(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -598,7 +593,7 @@ class Test_Friedrich_Dist2(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -635,7 +630,7 @@ class Test_Friedrich_Dist2(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -672,7 +667,7 @@ class Test_Friedrich_Dist2(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
@@ -709,7 +704,7 @@ class Test_Friedrich_Dist2(unittest.TestCase):
                                 error_limits=error_limits,
                                 units='',
                                 msg=title,
-                                output_filename=OUTPUT_FILE,
+                                output_dir=OUTPUT_LOC,
                                 create_output=CREATE_OUTPUT,
                                 verbose=VERBOSE)
 
